# Optimize stft - skip window multiplication for identity windows