#### Reference Issue
Fixes #986.

#### What does this implement/fix? Explain your changes.

This change skips window multiplication during stft when the window is the identity, thus avoids a possibly expensive array copy operation.


#### Any other comments?
Microbenchmarking is hard - I created a script so I can test out the speed improvement. I tried interleaving the new/old function calls to average out external factors.

This is the output:

```
Original:
Max: 0.09029326800000015s
Min: 0.04929946500000071s
Avg: 0.054472541203333356s

Optimized:
Max: 0.10773746899999992s
Min: 0.045211178999999824s
Avg: 0.04973186985666671s
Optimized wins 272/300 times (0.9066666666666666%)
```

On average the optimized version takes ~91.30% the time when tested with random 1000000 element arrays.

The execution times in a scatterplot, red are the original and blue the optimized:

![bench](https://user-images.githubusercontent.com/940823/65294315-3b36c580-db2c-11e9-9975-48283313dfed.png)


The correctness are covered by existing tests.