@lostanlen 2019-09-20
Thank you for starting this. Especially, many thanks for the thorough benchmark.

(1) please make sure to cover all synonyms of `ones`: `box`, `boxcar`, `ones`, `rect`, `rectangular`. 

(2) i wonder if this branch can be made even faster by `librosa.util.frame`?

(3) MATLAB regression tests are failing