@ZiyaoWei 2019-09-21
Thanks for the review! @lostanlen comments:

1) Didn't realize there are so many aliases, thanks!

2) Don't know the internal of librosa - will take a look!

3) Weird - I can't replicate this locally, and if you look at the numbers printed out in cI they seem to be the same - I might push a couple "debug" commits to this branch and try to figure out what's going on.