@lostanlen 2019-09-21
Re (2), forget what i said about `frame`. It's already framed, actually.

We'd have to check, but my gut feeling is that we don't necessarily need `bl_s` and `bl_t` in the transform. I would surmise that `fft.rfft` is clever enough to figure out how to optimally chunk this operation in order to make the most efficient use of CPU vectorization while avoiding cache overflow. This machine-specific optimization is the entire premise of the FFTW library. So, at least, it's worth trying a simplification of the branch down to

`stft_matrix = fft.rfft(y_frames, axis=0)`

and seeing how fast that is, and whether it produces out-of-memory errors.