@lostanlen 2019-09-24
Are you saying it's _faster_ to do a chunking heuristic before calling `fft.fft` rather than calling `fft.fft` over the entire matrix? Independently of FFT backend? Wow.