@lostanlen 2019-11-12
@ZiyaoWei hi! any updates regarding this PR? Last time we checked, you reported an acceleration from using a manual for loop rather than letting numpy do the vectorization, which sounded a bit surprising.

The acceleration is quite small for the moment. I think that an acceleration of at least 2x would be necessary to warrant making a special branch for this. Rectangular windows ("identity functions")  can be useful in bioacoustics for detecting sinusoids in noise, but they are almost never used in speech or music processing.