@bmcfee 2020-05-26
I just tried this out locally (independent reimplementation), and get similar behavior.  The block-wise acceleration is indeed faster.  I get about a 6ms boost by skipping the window multiply for the demo track (39ms down to 33ms average time).

I'm inclined to think this optimization isn't worth it.  But it's good to have done this exercise!