@lostanlen 2021-08-04
I am in favor of closing. `librosa.stft` already does the right thing and anyone wanting to try optimizing the boxcar STFT by skipping the window multiplication can obtain it very simply: (assuming `center=False`)

```python
def stft_rect(y, n_fft, hop_length, win_length, center=True, pad_mode="reflect"):
    if center:
        y = np.pad(y, int(n_fft // 2), mode=pad_mode)
    y_frames = librosa.util.frame(y, win_length, hop_length)
    fft = librosa.get_fftlib()
    stft_matrix = fft.rfft(y_frames, n=n_fft, axis=0)
```